requirements.txt
Django==2.1.7
psycopg2==2.7.7
djangorestframework==3.9.1
requests==2.21.0
dj_database_url
whitenoise
gunicorn
